refaftor: decouple Quiz and QuizLayout
