+++ resources/js/Pages/User/Quiz.vue
@@ -1,18 +1,14 @@
 <script setup lang="ts">
 import QuizLayout from '@/Layouts/QuizLayout.vue'
 import axios from 'axios'
-import FormButton from '@/components/Common/FormButton.vue'
-import Button from '@/components/Common/Button.vue'
 import { ref } from 'vue'
+import Button from '@/components/Common/Button.vue'
+import MessageBox from '@/components/Common/MessageBox.vue'
 
 defineProps<{ userQuiz: UserQuiz }>()
-
-const emptyAnswerMessage = ref(false)
 const networkErrorMessage = ref(false)
 
 function handleAnswer(question: UserQuestion, selectedAnswer: number) {
-  question.selectedAnswer = selectedAnswer
-
   axios.post(`/questions/${question.id}/${selectedAnswer}`, { _method: 'patch' }).catch(() => {
     networkErrorMessage.value = true
     question.selectedAnswer = undefined
@@ -23,46 +19,17 @@ function handleAnswer(question: UserQuestion, selectedAnswer: number) {
 <template>
   <QuizLayout 
     :user-quiz="userQuiz"
-    :empty-answer-message="emptyAnswerMessage"
-    :network-error-message="networkErrorMessage"
+    :request-close-quiz="{ method: 'post', href: `/quizzes/${userQuiz.id}/close` }"
     @answer="handleAnswer"
-  >
-    <template #submitButton="{ allQuestionsAnswered }">
-      <FormButton 
-        v-if="allQuestionsAnswered" 
-        large 
-        :href="`/quizzes/${userQuiz.id}/close`" 
-        method="post"
-      >
-        Oddaj test
-      </FormButton>
-      <Button 
-        v-else 
-        large 
-        @click="emptyAnswerMessage = true"
-      >
-        Oddaj test
-      </Button>
-    </template>
+  />
+  
+  <MessageBox :open="networkErrorMessage" @close="networkErrorMessage = false">
+    <template #title>Nie udało się wysłać odpowiedzi</template>
 
-    <template #submitWithoutAllAnswers>
-      <FormButton 
-        small 
-        :href="`/quizzes/${userQuiz.id}/close`" 
-        method="post"
-      >
-        Oddaj mimo to
-      </FormButton>
-    </template>
+    <template #message>Wystąpił problem z wysłaniem Twojej odpowiedzi. Sprawdź swoje połączenie internetowe i spróbuj ponownie.</template>
 
-    <template #timeoutButton>
-      <FormButton 
-        large 
-        :href="`/quizzes/${userQuiz.id}/result`" 
-        method="get"
-      >
-        Podsumowanie
-      </FormButton>
+    <template #buttons>
+      <Button small @click="networkErrorMessage = false">Ok</Button>
     </template>
-  </QuizLayout>
+  </MessageBox>
 </template>
